[WIP] More CUDA stuff (#26)
* initial commit


* LRN BP CUDA


* less memory


* Fixed bug with crop_and_resize op helper.

* get rid of unnecessary index-calculation dunction


* Fixed sort with nth_element cuda-based helper.

* Refactored nth_element.

* Refactored nth_element op and tests.

* Modified usage of dim array with sortTad routine.

* Refactored main routine of helper for non_max_image_suppression op.

* non_max_image_suppression op helper with cuda kernel implementation. Initial revision.

* fix vol2col cuda kernel

* meh


* topK concept


* unsorted topK with scanWitdh of 1


* correct vol2col tests

* sorted/unsorted topK


* implementation and fixing col2im/col2vol

* Corrected usage flags with input/output with reverse op.

* dup is const now


* percentile op


* group tests for mapool2d


* special test for george


* less threads for sortTad


* provide conv2d for cuda


* remove auther in sort tad kernel code


* provide depthwise_conv2d for cuda


* - max_pooling_with_argmax
- null check for special use


* dts cuda


* provide sconv2d for cuda


* std cuda


* Refactored non_max_suppression op to conform TF implementation.

* Improved suppression helper.

* provide pooling3d for cuda


* minor lstm rearrangements


* more of minor lstm rearrangements


* (bi)dynamic_rnn


* templates init order


* Refactored non_max_suppression op.

* Added cuda kernel for non_max_suppression.

* CPU sort by key/value


* CPU sort TAD by key/value


* CPU sort TAD by key/value tests


* Eliminate compiler error with cuda implementation.

* - repaired gradCheck in cuda
- provide conv2d_bp for cuda


* missed signature


* provide depthwise_conv2d_bp for cuda


* Implementation of lup helper with cuda kernel. Initial commit.

* further work on backprops for convolutions


* CUDA linear sort by key/val


* CUDA tad sort by key/val


* start providing of backprop for pooling2d/3d


* Added atomicAdd for bool datatype.

* dynamic partition concept


* dynamic partition concept


* dynamic partition scalar CUDA


* important comment


* fix pooling2d/3d backprop helpers


* Added non-linear test with dynamic_partition.

* Improved test for dynamic_partition.

* dynamic_partition TAD concept


* - dynamic_partition TAD CUDA impl
- dynamic_partition TAD CPU fix


* - rewrite cpu code for usampling2d/3d
- write cuda code for usampling2d/3d


* dynamic_stitch CUDA vector case


* dynamic_stitch CUDA TAD case concept


* dynamic_stitch CUDA TAD case impl


* Added tests for dynamic_stitch 3D-4D cases.

* minor tests tweaks


* Fixed type check for dynamic stitch.

* min/max bp


* rewrite code for upsampling2d/3d cpu


* reduce min/max/norm_max bp


* lup implementation. Additional enhancements.

* provide code for upsamling2d/3d backprop


* weightedCrossEntropyWithLogits


* Fixed template math atomicMul for 64bit ints.

* Refactored dynamic_partition_bp op.

* inverseBroadcast fix


* DynamicPartitionBP test datatype fixed.

* - nd4j_atomicMul Windows fix
- cpu/NDArrayLambda.hpp excluded from CUDA

+++ libnd4j/tests_cpu/layers_tests/DeclarableOpsTests9.cpp
@@ -2619,7 +2619,7 @@ TEST_F(DeclarableOpsTests9, Floormod_BP_Test_2) {
 TEST_F(DeclarableOpsTests9, Dynamic_Partition_BP_1) {
 
     auto x = NDArrayFactory::create<double>('c', {2, 3, 4});
-    auto y = NDArrayFactory::create<double>('c', {2, 3}, {0., 1., 2., 1., 0., 2. });
+    auto y = NDArrayFactory::create<int>('c', {2, 3}, {0, 1, 2, 1, 0, 2});
     auto dLdzX = NDArrayFactory::create<double>('c', {2, 4});
     auto dLdzY = NDArrayFactory::create<double>('c', {2, 4});
     auto dLdzZ = NDArrayFactory::create<double>('c', {2, 4});
